+++ notify/src/paths.rs
@@ -40,16 +40,12 @@ impl WatchPath {
 }
 
 impl WatchMetadata {
-    pub(crate) fn new<'a, I>(
+    pub(crate) fn new(
         path: &WatchPath,
         is_recursive: bool,
         is_user_watch: bool,
         existing_watch: Option<&Self>,
-        user_roots: I,
-    ) -> Self
-    where
-        I: IntoIterator<Item = (&'a PathBuf, &'a Self)>,
-    {
+    ) -> Self {
         let existing_reported_path = existing_watch.map(|watch| watch.reported_path.clone());
         let existing_is_user_watch = existing_watch.is_some_and(|watch| watch.is_user_watch);
         let existing_user_is_recursive =
@@ -61,18 +57,7 @@ impl WatchMetadata {
         } else if existing_is_user_watch {
             existing_reported_path.unwrap_or_else(|| path.requested.clone())
         } else {
-            user_roots
-                .into_iter()
-                .filter(|(candidate, watch)| {
-                    watch.is_user_watch
-                        && watch.user_is_recursive
-                        && path.absolute.starts_with(candidate)
-                })
-                .max_by_key(|(candidate, _)| candidate.as_os_str().len())
-                .map_or_else(
-                    || path.requested.clone(),
-                    |(root, watch)| reported_path(root, &watch.reported_path, &path.absolute),
-                )
+            path.requested.clone()
         };
 
         Self {